Везде в проекте ссылки указывают на доменные модели, id заменены на guid, url сервиса курса валют перемещен в конфигурационный файл и др. мелкие изменения

+++ Bank.Application/Clients/Commands/DeleteClient/DeleteClientCommandHandler.cs
@@ -1,4 +1,6 @@
-﻿using Bank.Application.Interfaces;
+﻿using Bank.Application.Common.Exeptions;
+using Bank.Application.Interfaces;
+using Bank.Domain.Client;
 using MediatR;
 
 namespace Bank.Application.Clients.Commands.DeleteClient;
@@ -11,9 +13,25 @@ public class DeleteClientCommandHandler : IRequestHandler<DeleteClientCommand>
     {
         _clientsDbContext = clientsDbContext;
     }
-    public async Task<Unit> Handle(DeleteClientCommand request, CancellationToken cancellationToken)
+    //public async Task<Unit> Handle(DeleteClientCommand request, CancellationToken cancellationToken)
+    //{
+    //    var targetClient = await _clientsDbContext.GetClient(request.Id, cancellationToken);
+    //    if (targetClient == null)
+    //    {
+    //        throw new NotFoundException(nameof(Client), request.Id);
+    //    }
+    //    await _clientsDbContext.DeleteClient(request.Id, cancellationToken);
+    //    return Unit.Value;
+    //}
+
+    public async Task Handle(DeleteClientCommand request, CancellationToken cancellationToken)
     {
+        var targetClient = await _clientsDbContext.GetClient(request.Id, cancellationToken);
+        if (targetClient == null)
+        {
+            throw new NotFoundException(nameof(Client), request.Id);
+        }
         await _clientsDbContext.DeleteClient(request.Id, cancellationToken);
-        return Unit.Value;
+        return;
     }
 }
